Women_Safety_Gesture_Recognisition-main/requirements.txt
streamlit
opencv-python
mediapipe
python-dotenv
pyrebase4
setuptools